Control Scene Done
Control Scene done. Just needs the pause menu to be implemented. Check out the pause script. It should be straight forward. Just needs to ensure that its active/listening for the 'y' pause button.

Im thinking that the pause scene needs to be a child of a specific node but I am unsure.

Currently has a print in the methods that are not being called. So that means its not listening at all. Once we get it to read keyboard inputs, it should work.